+++ main.cc
@@ -68,7 +68,7 @@ int main (int argc, char* argv[]) {
     sigprocmask(SIG_SETMASK, &orig_mask, NULL);
 
     n += v.size ();
-    for (unsigned int i = 0; i < v.size (); i++) {
+    for (size_t i = 0; i < v.size (); i++) {
       std::unique_ptr<evaristo> e = std::move(v[i]);
       if (!(c++ % prescale)) g.draw (e.get ());
       o.write (e.get ());
